MDL-67349 behat: Update setting select value event
Updated so that instead of triggering a change event, it will simply
click into the parent element.
Previously there were some assumptions about dialogues and the format
of the page, which could result in unexpected behaviour being triggered
(eg closing of a popover when setting a select element within it).